ToDotNot unit test conversion - part 11 (last).
