+++ src/Json.Schema.ToDotNet.UnitTests/TestData/DataModelGeneratorTests/GeneratesSealedClassesWhenOptionIsSet/ExpectedClass.cs
@@ -1,0 +1,12 @@
+using System;
+using System.CodeDom.Compiler;
+using System.Runtime.Serialization;
+
+namespace N
+{
+    [DataContract]
+    [GeneratedCode("Microsoft.Json.Schema.ToDotNet", "$JSchemaFileVersion$")]
+    public sealed class C
+    {
+    }
+}
