+++ src/Json.Schema.ToDotNet.UnitTests/TestData/DataModelGeneratorTests/GeneratesSealedClassesWhenOptionIsSet/Schema.json
@@ -1,0 +1,3 @@
+{
+  "type": "object",
+}
